examples/llm-teleprompter/requirements.txt
bleak>=0.21.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.18.0
ollama>=0.1.0
